USAboundaries: Historical and Contemporary Boundaries of the United States of America

This package provides contemporary (2014) boundaries for states, counties, zip code tabulation areas, and congressional districts in the United States of America. This data is provided by the U.S. Census Bureau.

Details

This package also provides spatial objects with historical boundaries of states or counties in the United States of America from 1629 to 2000. It provides data from the Atlas of Historical County Boundaries. The copyright to the historical data used in this package is owned by the Newberry Library, and it is included in the USAboundariesData package under the terms of the Creative Commons Attribution-NonCommercial-ShareAlike 2.5 Generic (CC BY-NC-SA 2.5) license.

The code in this package is copyrighted by Lincoln Mullen, and is released under the terms of the MIT License.