City locations and populations (historical and contemporary)

This function returns an sf object of cities (or populated places) with their populations and latitudes and longitudes. Population data is taken from the U.S. Census.

us_cities(map_date = NULL, states = NULL)

Arguments

map_date
If NULL, then city populations from the 2010 census (the most recent census) are returned. This parameter accepts a Date object or a character string coercible to a Date object, as well as numeric values representing a year. If a year or date is used, then city populations from the decennial census from 1790 to 2010 prior to that year is returned. For example, 1805 or "1805-07-04" would return city populations from the 1800 census.

states
A character vector of state or territory names or abbreviations. Only boundaries for those states/territories will be returned. If NULL, all boundaries will be returned.

References

The data was compiled by Erik Steiner and Jason Heppler at the Center for Spatial and Textual Analysis, Stanford University. See their description of the data for a fuller accounting of how the data was gathered.

Examples

if (require(USAboundariesData)) {
  us_cities(1805)
  us_cities("1828-05-08")
  us_cities()
}

#> old-style crs object detected; please recreate object with a recent sf::st_crs()

#> old-style crs object detected; please recreate object with a recent sf::st_crs()

#> old-style crs object detected; please recreate object with a recent sf::st_crs()

#> old-style crs object detected; please recreate object with a recent sf::st_crs()

#> City populations for contemporary data come from the 2010 census.

#> old-style crs object detected; please recreate object with a recent sf::st_crs()

#> old-style crs object detected; please recreate object with a recent sf::st_crs()

#> Simple feature collection with 6932 features and 12 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -166.5251 ymin: 19.09679 xmax: -67.23919 ymax: 71.25408
#> Geodetic CRS:  WGS 84
#> # A tibble: 6,932 x 13
#>    city   state_name state_abbr county    county_name  stplfips_2010 name_2010  
#>    <chr>  <chr>      <chr>      <chr>     <chr>        <chr>         <chr>      
#>  1 Ancho… Alaska     AK         THIRD JU… Third Judic… 0203000       Anchorage …
#>  2 Barrow Alaska     AK         North Sl… North Slope… 0205200       Barrow city
#>  3 Bethel Alaska     AK         Bethel C… Bethel Cens… 0206520       Bethel city
#>  4 Fairb… Alaska     AK         FOURTH J… Fourth Judi… 0224230       Fairbanks …
#>  5 Homer  Alaska     AK         Kenai Pe… Kenai Penin… 0233140       Homer city 
#>  6 Juneau Alaska     AK         FIRST JU… First Judic… 0236400       Juneau cit…
#>  7 Kenai  Alaska     AK         Kenai Pe… Kenai Penin… 0238420       Kenai city 
#>  8 Ketch… Alaska     AK         FIRST JU… First Judic… 0238970       Ketchikan …
#>  9 Kodiak Alaska     AK         Kodiak I… Kodiak Isla… 0240950       Kodiak city
#> 10 Kotze… Alaska     AK         Northwes… Northwest A… 0241830       Kotzebue c…
#> # … with 6,922 more rows, and 6 more variables: city_source <chr>,
#> #   population_source <chr>, place_type <chr>, year <int>, population <int>,
#> #   geometry <POINT [°]>