City locations and populations (historical and contemporary)

This function returns a data frame of cities (or populated places) with their populations and latitudes and longitudes. Population data is taken from the U.S. Census.

us_cities(map_date = NULL)

Arguments

map_date
If NULL, then city populations from the 2010 census (the most recent census) are returned. This parameter accepts a Date object or a character string coercible to a Date object, as well as numeric values representing a year. If a year or date is used, then city populations from the decennial census from 1790 to 2010 prior to that year is returned. For example, 1805 or "1805-07-04" would return city populations from the 1800 census.

References

The data was compiled by Erik Steiner and Jason Heppler at the Center for Spatial and Textual Analysis, Stanford University. See their description of the data for a fuller accounting of how the data was gathered.

Examples

if (require(USAboundariesData)) {
  us_cities(1805)
  us_cities("1828-05-08")
  us_cities()
}

#> City populations for contemporary data come from the 2010 census.

#> # A tibble: 6,933 x 15
#>         city state stplfips_2010               name_2010
#>        <chr> <chr>         <chr>                   <chr>
#>  1 Anchorage    AK       0203000  Anchorage municipality
#>  2    Barrow    AK       0205200             Barrow city
#>  3    Bethel    AK       0206520             Bethel city
#>  4 Fairbanks    AK       0224230          Fairbanks city
#>  5     Homer    AK       0233140              Homer city
#>  6    Juneau    AK       0236400 Juneau city and borough
#>  7     Kenai    AK       0238420              Kenai city
#>  8 Ketchikan    AK       0238970          Ketchikan city
#>  9    Kodiak    AK       0240950             Kodiak city
#> 10  Kotzebue    AK       0241830           Kotzebue city
#> # ... with 6,923 more rows, and 11 more variables: county <chr>, lat <dbl>,
#> #   lon <dbl>, lat_bing <dbl>, lon_bing <dbl>, `city source` <chr>, `population
#> #   source` <chr>, `place type` <chr>, county_name <chr>, year <int>,
#> #   population <int>